+++ Packages/com.holoi.reality.typography/Runtime/_Public/DefaultTypedRealityManager.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+using Unity.Netcode;
+using Holoi.Library.HoloKitApp;
+
+
+namespace Holoi.Reality.Typography
+{
+    public class DefaultTypedRealityManager : RealityManager
+    {
+
+    }
+}