+++ Packages/com.holoi.reality.typography/Runtime/_TypedRealityFactoryTest/DotSequenceController.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.VFX;
+
+namespace Holoi.Reality.Typography
+{
+    public class DotSequenceController : MonoBehaviour
+    {
+        public VisualEffect _vfx;
+
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}