+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Dragon health bar [@yuchen]
 
+- Added a reality for factory testing [@sizheng]
+
 ### Changed
 
 - Changed the audio of AceFire Basic Spell: Spawn [@sizheng]